backend/app/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-groq
langchain-openai
langchain-community
langgraph
tavily-python
